backend-python/requirements.txt
# FastAPI 框架
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# 数据库
sqlalchemy==2.0.25
aiosqlite==0.19.0

# 认证
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# 配置管理
python-dotenv==1.0.0

# 文档
swagger-ui-fastapi==0.0.1

# 开发工具
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
